# Add request statistic reporting for decoupled mode